requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==0.19.0
openai==1.3.0
google.generativeai==0.3.0
requests==2.26.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.97.0
email-validator==2.1.0
